feat(identityserver): implement core auth and registration features
- Added IdentityServer project and initial configuration
- Seeded data and created initial migration
- Implemented login and multi-step register page
- Configured client credentials flow for token requests
- Added custom profile service for claims handling
- Configured authentication and endpoints on the resource server

feat(gateway): setup gateway service with reverse proxy and auth

- Added Gateway service with reverse proxy configuration
- Integrated authentication into gateway config
- Tested gateway service with Postman
- Added remaining contracts and consumers to Auction and Search services
- Configured new client in Identity service for gateway integration

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,13 +51,14 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
             //TODO: Add current user as seller
 
-            auction.Seller = "teste";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -71,7 +73,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
-
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -80,7 +82,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
 
-            //TODO: Check if seller name matches username
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             auction.Item.Manufacturer = updateAuctionDto.Manufacturer ?? auction.Item.Manufacturer;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -97,13 +99,15 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
-        [HttpDelete]
+        [Authorize]
+        [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
 
             if (auction == null) return NotFound();
 
+            if (auction.Seller != User.Identity.Name) return Forbid();
             _context.Auctions.Remove(auction);
 
             await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString()});
